+++ lib/plex_assets.dart
@@ -1,0 +1,3 @@
+
+const String pdf = "packages/plex/assets/pdf.png";
+const String excel = "packages/plex/assets/excel.png";